+++ R/intensitynetUnd.R
@@ -13,6 +13,8 @@ MeanNodeIntensity.intensitynetUnd = function(obj, node_id){
   # If the intensity is already calculated, return it
   if(!is.null(igraph::vertex_attr(g, 'intensity', index=node_id))){
     if(!is.na(igraph::vertex_attr(g, "intensity", index=node_id))[1]){
+      
+      message("Warning: Node intensities were already calculated in a previous instance, returning the same intensity.")
       return(igraph::vertex_attr(g, 'intensity', index=node_id))
     } 
   }
@@ -27,8 +29,7 @@ MeanNodeIntensity.intensitynetUnd = function(obj, node_id){
     for (neighbor_id in neighbors_list){
       ev_mat[as.character(node_id), as.character(neighbor_id)] <- EdgeIntensity(obj = obj, 
                                                                                 node_id1 = igraph::V(g)[node_id]$name, 
-                                                                                node_id2 = igraph::V(g)[neighbor_id]$name,
-                                                                                z = obj$event_correction)
+                                                                                node_id2 = igraph::V(g)[neighbor_id]$name)
     }
     
     mean_intensity <- Reduce('+', ev_mat) / igraph::degree(g, node_id)
@@ -38,16 +39,17 @@ MeanNodeIntensity.intensitynetUnd = function(obj, node_id){
 }
 
 
-#' Calculates edgewise and mean nodewise intensities for Undirected networks
+#' Calculates edgewise and mean nodewise intensities for for Undirected networks and, for each edge, the proportions of
+#' all event covariates.
 #' 
-#' @name CalculateEventIntensities.intensitynetUnd
+#' @name RelateEventsToNetwork.intensitynetUnd
 #' 
 #' @param obj intensitynetUnd object
 #' 
-#' @return intensitynetUnd object with a graph containing all the intensities as attributes of its nodes and edges
+#' @return proper intensitynetUnd object with a graph containing the nodewise intensity in the node 
+#' attributes and the edgewise intensities and event covariate proportions as edge attributes.
 #' 
-#' @export
-CalculateEventIntensities.intensitynetUnd = function(obj){
+RelateEventsToNetwork.intensitynetUnd = function(obj){
   g <- obj$graph
   counts <- c()
   
@@ -56,13 +58,14 @@ CalculateEventIntensities.intensitynetUnd = function(obj){
     return(obj)
   }
   
-  tmp_obj <- AllEdgeIntensities.intensitynet(obj)
+  tmp_obj <- EdgeIntensitiesAndProportions.intensitynet(obj)
   g <- tmp_obj$graph
   
   pb = utils::txtProgressBar(min = 0, max = igraph::gorder(g), initial = 0) 
   message("Calculating node intensities...")
   
   # check if the intensities was previously calculated, if not, calculate them
+  v_count <- 0
   for(node_id in igraph::V(g)){
     
     utils::setTxtProgressBar(pb,node_id)
@@ -78,10 +81,17 @@ CalculateEventIntensities.intensitynetUnd = function(obj){
     }else if(is.na(igraph::vertex_attr(g, 'intensity', node_id))[1]){
       counts[[node_id]] <- 0
     }else{
+      v_count <- v_count + 1
       counts[[node_id]] <- igraph::vertex_attr(g, 'intensity', node_id)
     }
   }
   close(pb)
+  # If the intensity of all edges is already calculated return the object
+  if(v_count == length(igraph::V(g))){
+    message("Warning: Intensities were already calculated in a previous instance, returning the same object.")
+    return(obj)
+  } 
+  
   
   #g <- g %>% igraph::set_vertex_attr(name = "intensity", value = as.matrix(counts))
   g <- igraph::set_vertex_attr(g, name = "intensity", value = as.matrix(counts))