+++ NEWS.md
@@ -1,3 +1,15 @@
+# intensitynet 1.2
+
+* Added new functionality. Now the package can work with marked events (numerical, categorical or both). Each edge in the graph provided by the intensitynet object contains all the information of its associated events.  
+
+* Changed internal function name 'AllEdgeIntensities()' to 'EdgeIntensitiesAndProportions()'  In order to adapt its significate to the new functionalities.
+
+* Changed funcion name 'CalculateEventIntensities()' to 'RelateEventsToNetwork()'. In order to adapt its significate to the new functionalities.
+
+* Greatly reduced the computational time to calculate event-related intensities with the function 'RelateEventsToNetwork()'
+
+
+
 # intensitynet 1.1
 
 * Modified DESCRIPTION name references 'IntensityNet' to 'intensitynet' to match the package name.